Модулирован main.js
